requirements.txt

python-periphery
scikit-learn
opencv-python
matplotlib